debug

+++ f2_2.py
@@ -64,8 +64,8 @@ class F2_2(nn.Module):
 
     def forward(self, m0):
         T = 5e-4 * m0
-        xmm = np.round(self.xm / self.dlam + 64.5)
-        ymm = np.round(self.ym / self.dlam + 64.5)
+        xmm = np.round(self.xm / self.dlam + 64.5).astype(np.int)
+        ymm = np.round(self.ym / self.dlam + 64.5).astype(np.int)
         print(xmm)
         print(type(xmm))
         print(self.phy[xmm[0]][:,ymm[0]])
